Fixing error messages. separating DB methods to its own file.
